Implement string repetition functionality
The 'repeatString' function has been amended to take two parameters: 'str', representing the string to be repeated, and 'rep', representing the number of times the string should be repeated. For non-negative 'rep', the function will return the repeated string; if 'rep' is negative, the function will return "ERROR". Corresponding tests have been updated and are now all active.