Unskip and fix tests in leapYears.spec.js
Replaced all "test.skip" with "test" in leapYears.spec.js file to enable all the tests. Also, implemented the leapYears function logic to correctly analyze if a given year is a leap year or not. This will ensure that leap year calculations are correct and all relevant test cases are covered.